Move the import logic into methods on the ApiV2 class
